requirements.txt
uvicorn~=0.23.2
fastapi~=0.103.2
cassandra-driver~=3.28.0